+++ client/src/components/InsertForm.jsx
@@ -1,7 +1,8 @@
-/* eslint-disable react/prop-types */
 /* eslint-disable no-unused-vars */
+/* eslint-disable react/prop-types */
 import Tagify from '@yaireo/tagify';
 import { useEffect, useRef, useState } from "react";
+import Button from 'react-bootstrap/Button';
 import { useForm } from "react-hook-form";
 import { useNavigate } from 'react-router-dom';
 import api, { addPage } from "../api/api";
@@ -11,9 +12,124 @@ const Types = {
     nrd: "Non-research dissertation",
 }
 
+const styles = {
+    form:{
+        backgroundColor: "white",
+        width: "70%",
+        maxWidth: "3xl",
+        margin: "auto",
+        textAlign: "left",
+        display: "flex",
+        flexDirection: "column",
+        justifyContent: "center",
+        gap: "2.5rem", // Assuming 1 rem is equivalent to 2.5rem
+        padding: "1rem 2rem", // Assuming 1 rem is equivalent to 2.5rem
+        boxShadow: "0px 4px 8px rgba(0, 0, 0, 0.1)", // Example shadow values; adjust as needed
+    },
+    container:{
+        display: "flex",
+        flexDirection: "column",
+        gap: "1rem",
+    },
+    innerContainer:{
+        display: "flex",
+        flexDirection: "column",
+        gap: "1rem",
+        width: "40%", // Since "w-2/5" corresponds to a width of 2/5 or 40%
+    },    
+    label:{
+        fontSize: "1rem",
+        fontWeight: "600",
+        lineHeight: "1.75",
+        color: "#1a202c",
+    },
+    input: {
+        padding: "0.5rem",
+        width: "100%",
+        border: "none",
+        background: "transparent",
+        paddingTop: "0.375rem",
+        paddingLeft: "0.25rem",
+        color: "#1a202c",
+        placeholder: {
+            color: "#a0aec0",
+        },
+        focus: {
+            outline: "0",
+            ringOffsetWidth: "0",
+            ringWidth: "2px",
+            },
+        fontSize: "0.875rem",
+        lineHeight: "1.5",
+        borderRadius: "0.375rem",
+        boxShadow: "0 2px 4px rgba(0, 0, 0, 0.05)"
+    },
+    textarea: {
+        width: "100%",
+        borderRadius: "0.375rem", // Corresponds to rounded-md in Tailwind
+        borderWidth: "0",
+        paddingY: "0.375rem", // Corresponds to py-1.5 in Tailwind
+        color: "#1a202c", // Replace with your specific text color, here it's text-gray-900
+        boxShadow: "0 1px 2px 0 rgba(0, 0, 0, 0.05)", // Corresponds to shadow-sm in Tailwind
+        boxShadowFocus: "0 0 0 3px rgba(66, 153, 225, 0.5)", // Corresponds to focus:ring-2 in Tailwind
+        boxShadowFocusInset: "inset 0 0 0 1px #e2e8f0", // Corresponds to focus:ring-inset in Tailwind
+        boxShadowFocusColor: "#3490dc", // Corresponds to focus:ring-indigo-600 in Tailwind
+        lineHeight: "1.5", // Corresponds to sm:leading-6 in Tailwind
+        fontSize: "0.875rem", // Corresponds to sm:text-sm in Tailwind
+        placeholderColor: "#a0aec0", // Replace with your specific placeholder text color, here it's text-gray-400
+        focusOutline: "none", // Disable the default focus outline
+    },
+    select:{
+        width: "100%",
+        borderRadius: "0.375rem", // You may adjust the radius value according to your design
+        boxShadow: "0 1px 2px 0 rgba(0, 0, 0, 0.05)", // Adjust the shadow according to your design
+        borderWidth: "1px",
+        borderStyle: "solid",
+        borderColor: "#CBD5E0", // Adjust the color according to your design
+        ringWidth: "1px",
+        ringStyle: "inset",
+        ringColor: "#CBD5E0", // Adjust the color according to your design
+        focusRingWidth: "2px",
+        focusRingStyle: "inset",
+        focusRingColor: "#4F46E5", // Adjust the color according to your design
+        outline: "none",
+        padding: "0.75rem 1rem", // You may adjust the padding according to your design
+        color: "#4A5568", // Adjust the text color according to your design
+        lineHeight: "1.25", // You may adjust the line height according to your design
+    },
+    header:{
+        fontSize: "2.25rem", // Equivalent to text-4xl in Tailwind CSS
+        fontWeight: "bold",  // Equivalent to font-bold in Tailwind CSS
+        letterSpacing: "-0.01em", // Equivalent to tracking-tight in Tailwind CSS
+        color: "#1a202c", // Equivalent to text-gray-900 in Tailwind CSS
+    },
+    buttonContainer:{
+        marginTop: "1.5rem", // Assuming default font size is 1rem, as Tailwind uses 1.5rem for mt-6
+        display: "flex",
+        alignItems: "center",
+        justifyContent: "flex-end",
+        columnGap: "1.5rem", // Assuming default font size is 1rem, as Tailwind uses 1.5rem for gap-x-6
+    },
+    cancel:{
+        fontSize: "0.875rem",
+        fontWeight: "400",
+        lineHeight: "1.5",
+        color: "#1a202c",
+    },
+    add:{
+        backgroundColor: "#1a365d",
+        color: "#fff",
+        
+    }
+     
+
+}
+    
+    
+
 export default function InsertForm() {
     //TODO: IMPORTANT!! Put this states in App.jsx (they are for sure needed somewhere else later in development)
-    const [levels, setLevels] = useState(["Bachelor", "Master", "PhD"]);
+    const [levels, setLevels] = useState(["Bachelor", "Master"]);
     const [supervisors, setSupervisors] = useState([]);
     const [degrees, setDegrees] = useState([]);
     const [keywords, setKeywords] = useState([]);
@@ -22,18 +138,17 @@ export default function InsertForm() {
     const [serverError, setServerError] = useState(false);
     const [successfullySent, setSuccesfullySent] = useState(false);
 
-    const navigate = useNavigate();
+    const navigateTo = useNavigate();
 
     const { register, formState: { errors }, handleSubmit } = useForm()
     const inputRef = useRef(null);
-
 
 
     const onSubmit = (data) => {
         addPage({
             ...data,
             expiration: new Date(data.expiration).toISOString(),
-            supervisor: parseInt(data.supervisor),
+            supervisor: 1,
             coSupervisors: cosupervisors.map((cosupervisor) => cosupervisor.trim()), //TODO: Fix in database or here sending of cosupervisors
             keywords: keywords.map((keyword) => keyword.trim()),
             groups: [],
@@ -42,7 +157,7 @@ export default function InsertForm() {
             setSuccesfullySent(true);
             setTimeout(() => {
 
-                navigate('/');
+                navigateTo('/');
             }, 7500);
 
         })
@@ -75,11 +190,14 @@ export default function InsertForm() {
         <>
             <form
                 onSubmit={handleSubmit(onSubmit)}
-                className="bg-white w-full max-w-3xl mx-auto text-left flex flex-col gap-y-10 px-4 py-6 drop-shadow-lg"
+                style={styles.form}
             >
-                <h1 className="sans-serif text-4xl font-bold tracking-tight text-gray-900">New Proposal</h1>
-                <div className="flex flex-col gap-y-1">
-                    <label className="text-base font-semibold leading-7 text-gray-900" htmlFor='title'>
+                <h1 
+                className="sans-serif"
+                style={styles.header}
+                >New Proposal</h1>
+                <div style={styles.container}>
+                    <label style={styles.label} htmlFor='title'>
                         Title
                     </label>
                     {errors.title?.type === "required" && (
@@ -89,13 +207,13 @@ export default function InsertForm() {
                         {...register("title", { required: true })}
                         id="title"
                         autoComplete="title"
-                        className="p-2 w-full border-0 bg-transparent py-1.5 pl-1 text-gray-900 placeholder:text-gray-400 focus:ring-0 sm:text-sm sm:leading-6 rounded-md shadow-sm ring-offset-2 ring-2"
+                        style={styles.input}
                         placeholder="Title of the thesis"
                     />
                 </div>
 
-                <div className="flex flex-col gap-y-1">
-                    <label className="block text-sm font-medium leading-6 text-gray-900" htmlFor='description'>
+                <div style={styles.container}>
+                    <label style={styles.label} htmlFor='description'>
                         Description
                     </label>
                     {errors.description?.type === "required" && (
@@ -105,16 +223,26 @@ export default function InsertForm() {
                         {...register("description", { required: true })}
                         id="description"
                         rows={3}
-                        className="w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
+                        style={styles.textarea}
                         defaultValue={''}
                         placeholder="Short description about the thesis..."
                     />
 
                 </div>
 
-                <div className="flex justify-between items-center">
-                    <div className="flex flex-col gap-y-1 w-2/5">
-                        <label className=" text-sm font-medium leading-6 text-gray-900" htmlFor='expiration-date'>
+                <div 
+                style={{
+                    display: "flex",
+                    flexDirection: "row",
+                    justifyContent: "space-between",
+                    alignItems: "center",
+                }}
+                >
+                    <div style={styles.innerContainer}>
+                        <label 
+                        style={styles.label} 
+                        htmlFor='expiration-date'
+                        >
                             Expiration date
                         </label>
 
@@ -122,14 +250,14 @@ export default function InsertForm() {
                             {...register("expiration", { required: true })}
                             type='date'
                             id='expiration-date'
-                            className=" w-full rounded-md shadow-sm ring-1 ring-inset ring-gray-300 focus-within:ring-2 focus-within:ring-inset focus-within:ring-indigo-600 border border-gray-400 text-gray-700  leading-tight focus:outline-none focus:border-gray-500"
+                            style={styles.select}
                             min={new Date().toISOString().split('T')[0]}
                         />
 
                     </div>
 
-                    <div className="flex flex-col gap-y-1 w-2/5">
-                        <label className="text-sm font-medium leading-6 text-gray-900">
+                    <div style={styles.innerContainer}>
+                        <label style={styles.label}>
                             Level
                         </label>
                         {errors.level?.type === "required" && (
@@ -137,7 +265,7 @@ export default function InsertForm() {
                         )}
                         <select
                             {...register("level", { required: true })}
-                            className="w-full rounded-md shadow-sm ring-1 ring-inset ring-gray-300 focus-within:ring-2 focus-within:ring-inset focus-within:ring-indigo-600 border border-gray-400 text-gray-700  leading-tight focus:outline-none focus:border-gray-500"
+                            style={styles.select}
                             id="level"
                             name="level"
                         >
@@ -146,8 +274,8 @@ export default function InsertForm() {
                     </div>
                 </div>
 
-                <div className="flex flex-col gap-y-1">
-                    <label className="text-sm font-medium leading-6 text-gray-900">
+                <div style={styles.container}>
+                    <label style={styles.label}>
                         Programme/Degree
                     </label>
                     {errors.cds?.type === "required" && (
@@ -156,14 +284,14 @@ export default function InsertForm() {
                     <select
                         {...register("cds", { required: true })}
                         id="cds"
-                        className="p-2 w-full border-0 bg-transparent py-1.5 pl-1 text-gray-900 placeholder:text-gray-400 focus:ring-0 sm:text-sm sm:leading-6 rounded-md shadow-sm ring-offset-2 ring-2"
+                        style={styles.select}
                     >
                         {degrees.map(degree => <option key={degree.COD_DEGREE} value={degree.COD_DEGREE}>{degree.TITLE_DEGREE}</option>)}
                     </select>
                 </div>
 
-                <div className="flex flex-col gap-y-1">
-                    <label className="text-sm font-medium leading-6 text-gray-900">
+                <div style={styles.container}>
+                    <label style={styles.label}>
                         Type
                     </label>
                     {errors.type?.type === "required" && (
@@ -171,7 +299,7 @@ export default function InsertForm() {
                     )}
                     <select
                         {...register("type", { required: true })}
-                        className="p-2 w-full border-0 bg-transparent py-1.5 pl-1 text-gray-900 placeholder:text-gray-400 focus:ring-0 sm:text-sm sm:leading-6 rounded-md shadow-sm ring-offset-2 ring-2"
+                        style={styles.select}
                         id="type"
                         name="type"
                     >
@@ -185,8 +313,8 @@ export default function InsertForm() {
 
 
 
-                <div className="flex flex-col gap-y-1">
-                    <label className="text-sm font-medium leading-6 text-gray-900">
+                {/* <div style={styles.container}>
+                    <label style={styles.label}>
                         Supervisor
                     </label>
                     {errors.supervisor?.type === "required" && (
@@ -195,21 +323,21 @@ export default function InsertForm() {
                     <select
                         {...register("supervisor", { required: true })}
                         id="supervisor"
-                        className="p-2 w-full border-0 bg-transparent py-1.5 pl-1 text-gray-900 placeholder:text-gray-400 focus:ring-0 sm:text-sm sm:leading-6 rounded-md shadow-sm ring-offset-2 ring-2"
+                        style={styles.select}
                     >
                         {supervisors.map(supervisor => <option key={supervisor.id} value={supervisor.id}>{supervisor.name} {supervisor.surname}</option>)}
                     </select>
-                </div>
+                </div> */}
 
-                <div className="flex flex-col gap-y-1">
-                    <label className="block text-sm font-medium leading-6 text-gray-900">
+                <div style={styles.container}>
+                    <label style={styles.label}>
                         Co-Supervisors
                     </label>
                     <CosupervisorsInput setCosupervisors={setCosupervisors} />
                 </div>
 
-                <div className="flex flex-col gap-y-1">
-                    <label className="block text-sm font-medium leading-6 text-gray-900">
+                <div style={styles.container}>
+                    <label style={styles.label}>
                         Required knowledge
                     </label>
                     {errors.requiredKnowledge?.type === "required" && (
@@ -218,12 +346,12 @@ export default function InsertForm() {
                     <input
                         {...register("requiredKnowledge", { required: true })}
                         id="requiredKnowledge"
-                        className="p-2 w-full border-0 bg-transparent py-1.5 pl-1 text-gray-900 placeholder:text-gray-400 focus:ring-0 sm:text-sm sm:leading-6 rounded-md shadow-sm ring-offset-2 ring-2"
+                        style={styles.select}
                     />
                 </div>
 
-                <div className="flex flex-col gap-y-1">
-                    <label className="block text-sm font-medium leading-6 text-gray-900">
+                <div style={styles.container}>
+                    <label style={styles.label}>
                         Notes - <span className="italic font-normal">Optional</span>
                     </label>
 
@@ -231,14 +359,14 @@ export default function InsertForm() {
                         {...register("notes", { required: false })}
                         id="notes"
                         rows={4}
-                        className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
+                        style={styles.textarea}
                         defaultValue={''}
                         placeholder="Notes about the thesis..."
                     />
                 </div>
 
-                <div className="flex flex-col gap-y-1">
-                    <label className="block text-sm font-medium leading-6 text-gray-900">
+                <div style={styles.container}>
+                    <label style={styles.label}>
                         Keywords
                     </label>
                     {errors.keywords?.type === "required" && (
@@ -256,19 +384,22 @@ export default function InsertForm() {
                 )}
 
 
-                <div className="mt-6 flex items-center justify-end gap-x-6">
-                    <button
-                        type="button"
-                        onClick={() => navigate('/')}
-                        className="text-sm font-semibold leading-6 text-gray-900">
+                <div 
+                style={styles.buttonContainer}
+                >
+                    <Button 
+                    variant="outline-secondary"
+                    onClick={() => navigate('/')}
+                    >
                         Cancel
-                    </button>
-                    <button
+                    </Button>
+                    <Button
                         type="submit"
-                        className="rounded-md bg-blue-900 hover:bg-blue-800 px-3 py-2 text-sm font-semibold text-white shadow-sm focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
+                        variant='dark'
+                        style={styles.add}
                     >
                         Add
-                    </button>
+                    </Button>
                 </div>
             </form>
         </>
