+++ server/src/controllers/teachers.js
@@ -1,5 +1,4 @@
-const { PrismaClient } = require("@prisma/client");
-const prisma = new PrismaClient({log:["query"]});
+const prisma = require('./prisma');
 
 module.exports = {
   /**
